Handle grant types
